BigDecimal instead of double

+++ src/main/java/com/abc/account/type/AccountType.java
@@ -1,10 +1,12 @@
 package com.abc.account.type;
 
+import java.math.BigDecimal;
+
 /**
  * Created by aholub on 8/31/15.
  */
 public interface AccountType {
 
     String getName();
-    double interestEarned(double amount);
+    BigDecimal interestEarned(BigDecimal amount);
 }
